rename maui example
