+++ Integrations/MAUI/CadMauiApp/CadMauiApp.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31611.283
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MauiApp1", "MauiApp1.csproj", "{8165FA96-EC08-4DF1-A42D-0C85BD652A48}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CadMauiApp", "CadMauiApp.csproj", "{95A82BFE-B854-49D5-A317-C69971ED3580}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,12 +11,12 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{8165FA96-EC08-4DF1-A42D-0C85BD652A48}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8165FA96-EC08-4DF1-A42D-0C85BD652A48}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8165FA96-EC08-4DF1-A42D-0C85BD652A48}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
-		{8165FA96-EC08-4DF1-A42D-0C85BD652A48}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8165FA96-EC08-4DF1-A42D-0C85BD652A48}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8165FA96-EC08-4DF1-A42D-0C85BD652A48}.Release|Any CPU.Deploy.0 = Release|Any CPU
+		{95A82BFE-B854-49D5-A317-C69971ED3580}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{95A82BFE-B854-49D5-A317-C69971ED3580}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{95A82BFE-B854-49D5-A317-C69971ED3580}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
+		{95A82BFE-B854-49D5-A317-C69971ED3580}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{95A82BFE-B854-49D5-A317-C69971ED3580}.Release|Any CPU.Build.0 = Release|Any CPU
+		{95A82BFE-B854-49D5-A317-C69971ED3580}.Release|Any CPU.Deploy.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
