+++ Integrations/MAUI/CadMauiApp/Platforms/Android/MainApplication.cs
@@ -1,7 +1,7 @@
 ﻿using Android.App;
 using Android.Runtime;
 
-namespace MauiApp1;
+namespace CadMauiApp;
 
 [Application]
 public class MainApplication : MauiApplication
