+++ Integrations/MAUI/CadMauiApp/Platforms/Android/Resources/values/colors.xml
