+++ Integrations/MAUI/CadMauiApp/Platforms/MacCatalyst/AppDelegate.cs
@@ -1,6 +1,6 @@
 ﻿using Foundation;
 
-namespace MauiApp1;
+namespace CadMauiApp;
 
 [Register("AppDelegate")]
 public class AppDelegate : MauiUIApplicationDelegate
