+++ Integrations/MAUI/CadMauiApp/Platforms/MacCatalyst/Program.cs
@@ -1,7 +1,7 @@
 ﻿using ObjCRuntime;
 using UIKit;
 
-namespace MauiApp1;
+namespace CadMauiApp;
 
 public class Program
 {
