+++ Integrations/MAUI/CadMauiApp/Platforms/Tizen/Main.cs
@@ -2,7 +2,7 @@ using System;
 using Microsoft.Maui;
 using Microsoft.Maui.Hosting;
 
-namespace MauiApp1;
+namespace CadMauiApp;
 
 class Program : MauiApplication
 {
