+++ Integrations/MAUI/CadMauiApp/Platforms/Windows/App.xaml
@@ -1,8 +1,8 @@
 ﻿<maui:MauiWinUIApplication
-    x:Class="MauiApp1.WinUI.App"
+    x:Class="CadMauiApp.WinUI.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:maui="using:Microsoft.Maui"
-    xmlns:local="using:MauiApp1.WinUI">
+    xmlns:local="using:CadMauiApp.WinUI">
 
 </maui:MauiWinUIApplication>
