+++ Integrations/MAUI/CadMauiApp/Platforms/Windows/Package.appxmanifest
@@ -8,7 +8,7 @@
 
   <Identity Name="maui-package-name-placeholder" Publisher="CN=User Name" Version="0.0.0.0" />
 
-  <mp:PhoneIdentity PhoneProductId="4B6914D6-4BC5-4D83-9A77-9A93159C6A8B" PhonePublisherId="00000000-0000-0000-0000-000000000000"/>
+  <mp:PhoneIdentity PhoneProductId="6D64C926-B3B3-4512-B1A7-58C73A95330E" PhonePublisherId="00000000-0000-0000-0000-000000000000"/>
 
   <Properties>
     <DisplayName>$placeholder$</DisplayName>
