+++ Integrations/MAUI/CadMauiApp/Resources/Fonts/OpenSans-Regular.ttf
