+++ Integrations/MAUI/CadMauiApp/Resources/Fonts/OpenSans-Semibold.ttf
