+++ Integrations/MAUI/CadMauiApp/Resources/Raw/AboutAssets.txt
