+++ Integrations/MAUI/CadMauiApp/Resources/Styles/Colors.xaml
