+++ Integrations/MAUI/CadMauiApp/Resources/Styles/Styles.xaml
